% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_trial_design.R
\name{add_design_layout}
\alias{add_design_layout}
\title{Specify design layout}
\usage{
add_design_layout(plot_length, field_col)
}
\arguments{
\item{plot_length}{numeric}

\item{field_col}{numeric}
}
\value{
data.frame
}
\description{
Specify design layout
}
