% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_field_parameters.R
\name{add_errors_to_par}
\alias{add_errors_to_par}
\title{Add measurement errors to a variable}
\usage{
add_errors_to_par(
  par_name,
  scaler,
  new_var_name,
  cell_data,
  nsim,
  xy,
  sp_range,
  gstat_model
)
}
\arguments{
\item{par_name}{(character)}

\item{scaler}{(data.frame)}

\item{new_var_name}{(character)}

\item{cell_data}{(data.frame)}

\item{nsim}{(numeric)}

\item{xy}{(sf)}

\item{sp_range}{numeric}

\item{gstat_model}{(character)}
}
\value{
data.frame of input rate, block_id, and plot_id
}
\description{
Add measurement errors to a variable
}
