% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_trial_design.R
\name{assign_input_rate}
\alias{assign_input_rate}
\title{Define the levels of experimental input rate}
\usage{
assign_input_rate(N_levels, block_num, design)
}
\arguments{
\item{N_levels}{(numeric)}

\item{block_num}{(numeric)}

\item{design}{(character)}
}
\value{
data.frame of input rate, block_id, and plot_id
}
\description{
Define the levels of experimental input rate by simulation id
}
