% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_trial_design.R
\name{assign_trial_design}
\alias{assign_trial_design}
\title{Delineate plot and treatment blocks on a field}
\usage{
assign_trial_design(field_data, design = "Latin Square Fixed 5")
}
\arguments{
\item{field_data}{data.frame (object generated by gen_fields())}

\item{design}{character (design name)}
}
\value{
data.frame
}
\description{
Delineate plot and treatment blocks on a field. The plot and block ids are added as columns to field_sf of the data generated by gen_fields()
}
