% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_field_parameters.R
\name{gen_field_parameters}
\alias{gen_field_parameters}
\title{Generate cell-level field parameters}
\usage{
gen_field_parameters(field_sf, sp_range, gstat_model, nsim)
}
\arguments{
\item{field_sf}{(sf)}

\item{sp_range}{(numeric)}

\item{gstat_model}{(character)}

\item{nsim}{numeric}
}
\value{
data.frame of input rate, block_id, and plot_id
}
\description{
Generate cell-level field parameters
}
