% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_field_parameters.R
\name{gen_field_pars}
\alias{gen_field_pars}
\title{## Supporting functions
## Splitting parameters
### split b0
### Split plateau
### Split Nk
### Split b1
### Split b2
## Create irrelevant parameters}
\usage{
gen_field_pars(sp_range, gstat_model, field_sf, nsim)
}
\description{
## Supporting functions
## Splitting parameters
### split b0
### Split plateau
### Split Nk
### Split b1
### Split b2
## Create irrelevant parameters
}
