% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_fields.R
\name{gen_fields}
\alias{gen_fields}
\title{Generate a field}
\usage{
gen_fields(
  plot_length = 12,
  plot_width = 3,
  cell_buffer = 1,
  aunit_length = 2,
  aunit_width = 3,
  cell = 6,
  field_col = 144,
  field_row = 72
)
}
\arguments{
\item{plot_length}{= 12 # the length of a plot (in number of cells)}

\item{plot_width}{= 3 # the width of a plot (in number of cells)}

\item{cell_buffer}{= 1}

\item{aunit_length}{= 2 # the length of an analysis unit (in number of cells)}

\item{aunit_width}{= 3 # the width of an analysis unit (in number of cells)}

\item{cell}{= 6 # the length of a cell in meter}

\item{field_col}{= 144 # the number of cell columns (determine how wide the field is)}

\item{field_row}{= 72, # the number of row columns (determine how tall the field is)}
}
\value{
data.frame where field_sf column holds the created field as an sf object
}
\description{
Generate a rectangular field on which on-farm precision experiments are conducted.
}
