% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_field_parameters.R
\name{gen_pars}
\alias{gen_pars}
\title{Generate parameters}
\usage{
gen_pars(mean, psill, range, coef_name, gstat_model, xy, nsim)
}
\arguments{
\item{mean}{(character)}

\item{psill}{(data.frame)}

\item{range}{(numeric)}

\item{coef_name}{(character)}

\item{gstat_model}{(data.frame)}

\item{xy}{(sf)}

\item{nsim}{(numeric)}
}
\value{
data.frame
}
\description{
Generate parameters in a spatially correlated manner using random gaussian process
}
