% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_trial_design.R
\name{gen_plot_block_ids}
\alias{gen_plot_block_ids}
\title{Create plot and block ids based on the design layout}
\usage{
gen_plot_block_ids(
  field_sf,
  plot_length,
  plot_width,
  cols_plot_in_block,
  rows_plot_in_block,
  cell_buffer
)
}
\arguments{
\item{field_sf}{data.frame (object generated by gen_fields())}

\item{plot_length}{numeric}

\item{plot_width}{numeric}

\item{cols_plot_in_block}{numeric}

\item{rows_plot_in_block}{numeric}

\item{cell_buffer}{numeric}

\item{design}{character (design name)}
}
\value{
data.frame
}
\description{
Create plot and block ids based on the design. Used internally.
}
