% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_weight_matrix.R
\name{gen_weight_matrix}
\alias{gen_weight_matrix}
\title{Generate distance-based weight matrix}
\usage{
gen_weight_matrix(reg_data, cutoff)
}
\arguments{
\item{reg_data}{(data.frame)}

\item{cutoff}{(numeric)}
}
\value{
A matrix of weights
}
\description{
Generate distance-based weight matrix
}
