% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_analysis_data.R
\name{gen_yield_QD}
\alias{gen_yield_QD}
\title{Generate yield}
\usage{
gen_yield_QD(b0, b1, b2, N)
}
\arguments{
\item{b0}{(numeric)}

\item{b1}{(numeric)}

\item{b2}{(numeric)}

\item{N}{(numeric)}

\item{Nk}{(numeric)}
}
\value{
(numeric) yield values
}
\description{
Generate yield according to quadratic function
}
