% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_analysis_data.R
\name{gen_yield_QP}
\alias{gen_yield_QP}
\title{Generate yield}
\usage{
gen_yield_QP(b0, b1, b2, Nk, N)
}
\arguments{
\item{b0}{(numeric)}

\item{b1}{(numeric)}

\item{b2}{(numeric)}

\item{Nk}{(numeric)}

\item{N}{(numeric)}
}
\value{
(numeric) yield values
}
\description{
Generate yield according to quadratic-plateau function
}
