% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_fields.R
\name{make_field}
\alias{make_field}
\title{Create a field}
\usage{
make_field(field_col, field_row, aunit_length, aunit_width, cell, cell_buffer)
}
\arguments{
\item{field_col}{= 144 # the number of cell columns (determine how wide the field is)}

\item{field_row}{= 72, # the number of row columns (determine how tall the field is)}

\item{aunit_length}{= 2 # the length of an analysis unit (in number of cells)}

\item{aunit_width}{= 3 # the width of an analysis unit (in number of cells)}

\item{cell}{= 6 # the length of a cell in meter}

\item{cell_buffer}{= 1}
}
\value{
sf object of the created field
}
\description{
Generate a rectangular field on which on-farm precision experiments are conducted. A utility function used internally by gen_fields()
}
