% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_field_parameters.R
\name{split_par_min}
\alias{split_par_min}
\title{Split a variable in an additive manner (internal)}
\usage{
split_par_min(par_name, cell_data, nsim, xy, sp_range, gstat_model)
}
\arguments{
\item{par_name}{(character)}

\item{cell_data}{(data.frame)}

\item{nsim}{(character)}

\item{xy}{(sf)}

\item{sp_range}{numeric}

\item{gstat_model}{(character)}
}
\value{
data.frame of input rate, block_id, and plot_id
}
\description{
Split a variable in an additive manner (internal)
}
