% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_field_parameters.R
\name{gen_field_parameters}
\alias{gen_field_parameters}
\title{# Generate field parameters for quadratic plateau production function}
\usage{
gen_field_parameters(field_with_design, nsim = 1000)
}
\description{
# Generate field parameters for quadratic plateau production function
}
